+++ library/Dlayer/DesignerTool/ContentManager/BlogPost/SubTool/Styling/Form.php
@@ -6,7 +6,7 @@
  * @license https://github.com/Dlayer/dlayer/blob/master/LICENSE
  */
 class Dlayer_DesignerTool_ContentManager_BlogPost_SubTool_Styling_Form extends
-    Dlayer_DesignerTool_ContentManager_Shared_Form_Styling
+    Dlayer_DesignerTool_ContentManager_Shared_Form_Content_Styling
 {
     /**
      * Set the properties for the form