+++ library/Dlayer/DesignerTool/ContentManager/Heading/Tool.php
@@ -73,7 +73,7 @@ class Dlayer_DesignerTool_ContentManager_Heading_Tool extends Dlayer_Tool_Conten
 	 */
 	protected function validateInstances($site_id, $content_id)
 	{
-        $model = new Dlayer_DesignerTool_ContentManager_Shared_Model_Content();
+        $model = new Dlayer_DesignerTool_ContentManager_Shared_Model_Content_Item();
         $instances = $model->instancesOfData($site_id, $content_id, 'Heading');
 
 		if($instances > 1)