+++ library/Dlayer/DesignerTool/ContentManager/HeadingDate/Tool.php
@@ -94,7 +94,7 @@ class Dlayer_DesignerTool_ContentManager_HeadingDate_Tool extends Dlayer_Tool_Co
      */
     protected function validateInstances($site_id, $content_id)
     {
-        $model = new Dlayer_DesignerTool_ContentManager_Shared_Model_Content();
+        $model = new Dlayer_DesignerTool_ContentManager_Shared_Model_Content_Item();
         $instances = $model->instancesOfData($site_id, $content_id, 'HeadingDate');
 
         if ($instances > 1) {