+++ library/Dlayer/DesignerTool/ContentManager/RichText/SubTool/Typography/Tool.php
@@ -6,7 +6,7 @@
  * @license https://github.com/Dlayer/dlayer/blob/master/LICENSE
  */
 class Dlayer_DesignerTool_ContentManager_RichText_SubTool_Typography_Tool extends
-    Dlayer_DesignerTool_ContentManager_Shared_Tool_Typography
+    Dlayer_DesignerTool_ContentManager_Shared_Tool_Content_Typography
 {
     /**
      * Generate the return ids array