+++ library/Dlayer/DesignerTool/ContentManager/Shared/Form/Content/Styling.php
@@ -7,7 +7,7 @@
  * @copyright G3D Development Limited
  * @license https://github.com/Dlayer/dlayer/blob/master/LICENSE
  */
-class Dlayer_DesignerTool_ContentManager_Shared_Form_Styling extends Dlayer_Form_Tool_Content
+class Dlayer_DesignerTool_ContentManager_Shared_Form_Content_Styling extends Dlayer_Form_Tool_Content
 {
     /**
      * Set the properties for the form