+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 
 Full changelog for Dlayer since the v1.00 release.
 
+## v1.17: 
+
+* New styling table for content items, stores all defined styles, no need for a table per styling group. [Refactoring]
+* Cleaned up tools, removed redundant model classes and moved around shared classes, now more apparent if they relate to content or not. [Refactoring]
+
 ## v1.16: 12th March 2017
 
 * Enabled link support in Quill rich text fields. [v0.30.0 of PHP Quill renderer]